Add a new custom keygenerator
The new generator will create keys with format
<computation_time>:<result_size>. To use the custom protocol enable the
preprocessing macro USE_CUSTOM_PROTOCOL in SConstruct.

The custom key generator takes as arguments two other Generators, one
for the computation time and one for the value size. For the computation
time we use the keysize generator, as usual with -K, for the value size
we use the valuesize generator, as usual with -V. This way we can create
exponential service times.